Replace ValidateTaskProperties with ValidatePlugin
