+++ subprojects/distributions/src/changes/accepted-public-api-changes.json
@@ -115,6 +115,16 @@
             "changes": [
                 "org.gradle.api.artifacts.repositories.UrlArtifactRepository"
             ]
+        },
+        {
+            "type": "org.gradle.plugin.devel.tasks.ValidateTaskProperties",
+            "member": "Class org.gradle.plugin.devel.tasks.ValidateTaskProperties",
+            "acceptation": "ValidateTaskProperties is no longer a ConventionTask",
+            "changes": [
+                "org.gradle.api.internal.ConventionTask.conventionMapping(java.lang.String,groovy.lang.Closure)",
+                "org.gradle.api.internal.ConventionTask.conventionMapping(java.lang.String,java.util.concurrent.Callable)",
+                "org.gradle.api.internal.ConventionTask.getConventionMapping()"
+            ]
         }
     ]
 }
