+++ subprojects/docs/src/docs/userguide/upgrading_version_5.adoc
@@ -105,12 +105,15 @@ Custom logic can be used from a settings script by <<tutorial_using_tasks.adoc#s
 
 The `org.gradle.api.Nullable` annotation type has been removed. Use `javax.annotation.Nullable` from JSR-305 instead.
 
-==== Task validation changes
+==== Plugin validation changes
 
-- `ValidateTaskProperties.setClasses()` is now removed. Use `getClasses().setFrom()` instead.
-- `ValidateTaskProperties.setClasspath()` is now removed. use `getClasspath().setFrom()` instead.
-- link:{javadocPath}/org/gradle/plugin/devel/tasks/ValidateTaskProperties.html#getFailOnWarning--[ValidateTaskProperties.failOnWarning] is enabled by default.
-- The following task validation errors now fail the build at runtime and are promoted to errors for link:{javadocPath}/org/gradle/plugin/devel/tasks/ValidateTaskProperties.html[ValidateTaskProperties]:
+- The `validateTaskProperties` task is now deprecated, use `validatePlugin` instead.
+  The new name better reflects the fact that it also validates artifact transform parameters and other non-property definitions.
+- The `ValidateTaskProperties` type is replaced by `ValidatePlugin`.
+- The `setClasses()` method is now removed. Use `getClasses().setFrom()` instead.
+- The `setClasspath()` method is also removed. use `getClasspath().setFrom()` instead.
+- The link:{javadocPath}/org/gradle/plugin/devel/tasks/ValidatePlugin.html#getFailOnWarning--[failOnWarning] option is now enabled by default.
+- The following task validation errors now fail the build at runtime and are promoted to errors for link:{javadocPath}/org/gradle/plugin/devel/tasks/ValidatePlugin.html[ValidatePlugin]:
   * A task property is annotated with a property annotation not allowed for tasks, like `@InputArtifact`.
 
 ==== `DefaultTask` and `ProjectLayout` methods replaced with `ObjectFactory`
