+++ subprojects/plugin-development/src/integTest/groovy/org/gradle/plugin/devel/tasks/ValidatePluginConfiguredByPluginIntegrationTest.groovy
@@ -17,9 +17,8 @@
 package org.gradle.plugin.devel.tasks
 
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
-import spock.lang.Unroll
 
-class ValidateTaskPropertiesConfiguredByPluginIntegrationTest extends AbstractIntegrationSpec {
+class ValidatePluginConfiguredByPluginIntegrationTest extends AbstractIntegrationSpec {
 
     def setup() {
         buildFile << """
@@ -27,42 +26,9 @@ class ValidateTaskPropertiesConfiguredByPluginIntegrationTest extends AbstractIn
         """
     }
 
-    @Unroll
-    def "delegates to validatePlugin and emits deprecation warning when configuration is changed via #methodCall for validateTaskProperties"() {
+    def "detects missing annotations on Java properties"() {
         buildFile << """
-            validateTaskProperties {
-                ${methodCall}
-            }
-
-            ${check ? "assert validatePlugin." + check : ""}
-        """
-
-        executer.expectDeprecationWarning()
-
-        when:
-        run "help"
-
-        then:
-        output.contains("The validateTaskProperties task has been deprecated. This is scheduled to be removed in Gradle 7.0. Please use the validatePlugin task instead.")
-
-        where:
-        methodCall                        | check
-        "getIgnoreFailures()"             | null
-        "ignoreFailures = true"           | "ignoreFailures == true"
-        "ignoreFailures = false"          | "ignoreFailures == false"
-        "getFailOnWarning()"              | null
-        "failOnWarning = true"            | "failOnWarning == true"
-        "failOnWarning = false"           | "failOnWarning == false"
-        "getClasses()"                    | null
-        "getClasspath()"                  | null
-        "getEnableStricterValidation()"   | null
-        "enableStricterValidation = true" | "enableStricterValidation == true"
-        "getOutputFile()"                 | null
-    }
-
-    def "detects missing annotations on Java properties while emitting deprecation warning"() {
-        buildFile << """
-            validateTaskProperties {
+            validatePlugin {
                 failOnWarning = true
             }
         """
@@ -141,10 +107,8 @@ class ValidateTaskPropertiesConfiguredByPluginIntegrationTest extends AbstractIn
             }
         """
 
-        executer.expectDeprecationWarning()
-
         expect:
-        fails "validateTaskProperties"
+        fails "check"
         failure.assertHasCause "Task property validation failed"
         failure.assertHasCause "Warning: Type 'com.example.MyTask': property 'badTime' is not annotated with an input or output annotation."
         failure.assertHasCause "Warning: Type 'com.example.MyTask': property 'options.badNested' is not annotated with an input or output annotation."
@@ -155,6 +119,5 @@ class ValidateTaskPropertiesConfiguredByPluginIntegrationTest extends AbstractIn
             Warning: Type 'com.example.MyTask': property 'options.badNested' is not annotated with an input or output annotation.
             Warning: Type 'com.example.MyTask': property 'ter' is not annotated with an input or output annotation.
             """.stripIndent().trim()
-        output.contains "The validateTaskProperties task has been deprecated. This is scheduled to be removed in Gradle 7.0. Please use the validatePlugin task instead."
     }
 }
